halfway through moving stuff out of forest module into forest_helper module (switching dev containers)
