+++ ws3/common.py
@@ -1123,3 +1123,5 @@ class Tree:
     #            graph.edge(str(n.nid), str(c.nid))
     #    graph.graph_attr.update(size='10,10')
     #    return graph
+
+
